Major final changes that get weapon firing and turn rotations going.
Lots of changes but the essential gameplay framework is now implemented.
Each player gets to fire their weapon, and hitting any players fort (including
your own) does damage to the fort.  After the fort has lost all of its health, the
player is flagged dead and they no longer get a turn.  This continues until only one, or no players, remain, but the game just stops at that point.

There is still a major bug with the RenderQueue that causes the program to crash
when removing the weapon projectiles and forts, but the are removed from the collision
checking part of the scene, so shots will pass through destroyed forts.  I can't figure
out what is going on, since the weapon cursor removes from the renderQueue without
issue.